requirements.txt
transformers>=4.40.0
datasets>=2.18.0
scikit-learn>=1.3.0
torch>=2.1.0
matplotlib>=3.8.0
fastapi
uvicorn
sentence-transformers
faiss-cpu
tqdm
openai
rouge-score
bert-score
detoxify
